[fix] removed the entirety of edit-template, everything now takes place in app/template
